requirements.txt
pandas
Pillow
svgwrite
svgpathtools
numpy
opencv-python
tensorflow==1.13.1
streamlit
streamlit-drawable-canvas==0.9.2
